% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pagespeed_as_dataframe}
\alias{pagespeed_as_dataframe}
\title{Convert a Google Pagespeed Object to a dataframe}
\usage{
pagespeed_as_dataframe(pagespeed_api)
}
\arguments{
\item{pagespeed_api}{A object of class 'pagespeed_api';
normally returned from `get_pagespeed_api`.}
}
\value{
A dataframe
}
\description{
Convert a Google Pagespeed Object to a dataframe
}
\examples{
\dontrun{
pagespeed_as_dataframe(get_pagespeed_api("http://finovate.com/blog/", "1234"))
}
}

